+++ sdk/resourcemanager/Proto.Client/src/Scenarios/AddTagToGeneric.cs
@@ -12,8 +12,8 @@ namespace Proto.Client
             var createVm = new CreateSingleVmExample(Context);
             createVm.Execute();
 
-            var rgOp = new AzureResourceManagerClient(new DefaultAzureCredential()).GetResourceGroupOperations(Context.SubscriptionId, Context.RgName);
-            foreach (var genericOp in rgOp.GetVirtualMachineContainer().ListAsGenericResource(Context.VmName))
+            var rgOp = new ArmClient(new DefaultAzureCredential()).GetResourceGroupOperations(Context.SubscriptionId, Context.RgName);
+            foreach (var genericOp in rgOp.GetVirtualMachines().ListAsGenericResource(Context.VmName))
             {
                 Console.WriteLine($"Adding tag to {genericOp.Id}");
                 genericOp.StartAddTag("tagKey", "tagVaue");