+++ sdk/resourcemanager/Proto.Client/src/Scenarios/TenantResource.cs
@@ -10,7 +10,7 @@ namespace Proto.Client
     {
         public override void Execute()
         {
-            var client = new AzureResourceManagerClient(new DefaultAzureCredential());
+            var client = new ArmClient(new DefaultAzureCredential());
             TenantOperations tenant = client.Tenant;
             string testAccount = "9b84fb4c-c79a-4819-bd24-1e58f88885bd";
             // old value: "3984c6f4-2d2a-4b04-93ce-43cf4824b698:e2f1492a-a492-468d-909f-bf7fe6662c01_2019-05-31";