+++ sdk/servicebus/Azure.Messaging.ServiceBus/tests/Client/ServiceBusClientTests.cs
@@ -20,7 +20,7 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         ///
         public static IEnumerable<object[]> ConstructorTokenCredentialArgumentInvalidCases()
         {
-            var credential = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>(), "{namespace}.servicebus.windows.net");
+            var credential = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>());
 
             yield return new object[] { null, credential.Object };
             yield return new object[] { "", credential.Object };
@@ -34,7 +34,21 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         ///
         public static IEnumerable<object[]> ConstructorSharedKeyCredentialArgumentInvalidCases()
         {
-            var credential = new ServiceBusSharedAccessKeyCredential("name", "value");
+            var credential = new AzureNamedKeyCredential("name", "value");
+
+            yield return new object[] { null, credential };
+            yield return new object[] { "", credential };
+            yield return new object[] { "FakeNamespace", null };
+            yield return new object[] { "sb://fakenamspace.com", credential };
+        }
+
+        /// <summary>
+        ///   Provides the invalid test cases for the constructor tests.
+        /// </summary>
+        ///
+        public static IEnumerable<object[]> ConstructorSasCredentialArgumentInvalidCases()
+        {
+            var credential = new AzureSasCredential(new SharedAccessSignature("amqps://fake.namespace.com", "name", "value").Value);
 
             yield return new object[] { null, credential };
             yield return new object[] { "", credential };
@@ -48,7 +62,7 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         ///
         public static IEnumerable<object[]> ConstructorCreatesDefaultOptionsCases()
         {
-            var credential = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>(), "{namespace}.servicebus.windows.net");
+            var credential = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>());
             var fakeConnection = "Endpoint=sb://not-real.servicebus.windows.net/;SharedAccessKeyName=DummyKey;SharedAccessKey=[not_real];EntityPath=fake";
 
             yield return new object[] { new ReadableOptionsMock(fakeConnection), "simple connection string" };
@@ -62,7 +76,7 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         ///
         public static IEnumerable<object[]> ConstructorClonesOptionsCases()
         {
-            var credential = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>(), "{namespace}.servicebus.windows.net");
+            var credential = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>());
             var fakeConnection = "Endpoint=sb://not-real.servicebus.windows.net/;SharedAccessKeyName=DummyKey;SharedAccessKey=[not_real];EntityPath=fake";
 
             var options = new ServiceBusClientOptions
@@ -176,9 +190,22 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         [Test]
         [TestCaseSource(nameof(ConstructorSharedKeyCredentialArgumentInvalidCases))]
         public void ConstructorValidatesSharedKeyArguments(string fullyQualifiedNamespace,
-                                                             object credential)
+                                                           object credential)
         {
-            Assert.That(() => new ServiceBusClient(fullyQualifiedNamespace, (ServiceBusSharedAccessKeyCredential)credential), Throws.InstanceOf<ArgumentException>());
+            Assert.That(() => new ServiceBusClient(fullyQualifiedNamespace, (AzureNamedKeyCredential)credential), Throws.InstanceOf<ArgumentException>());
+        }
+
+        /// <summary>
+        ///    Verifies functionality of the <see cref="ServiceBusClient" />
+        ///    constructor.
+        /// </summary>
+        ///
+        [Test]
+        [TestCaseSource(nameof(ConstructorSasCredentialArgumentInvalidCases))]
+        public void ConstructorValidatesSasArguments(string fullyQualifiedNamespace,
+                                                     object credential)
+        {
+            Assert.That(() => new ServiceBusClient(fullyQualifiedNamespace, (AzureSasCredential)credential), Throws.InstanceOf<ArgumentException>());
         }
 
         /// <summary>
@@ -241,7 +268,7 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         [Test]
         public void ConstructorWithTokenCredentialArgumentsValidatesOptions()
         {
-            var token = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>(), "{namespace}.servicebus.windows.net");
+            var token = new Mock<ServiceBusTokenCredential>(Mock.Of<TokenCredential>());
             var invalidOptions = new ServiceBusClientOptions { TransportType = ServiceBusTransportType.AmqpTcp, WebProxy = Mock.Of<IWebProxy>() };
             Assert.That(() => new ServiceBusClient("fullyQualifiedNamespace", Mock.Of<TokenCredential>(), invalidOptions), Throws.InstanceOf<ArgumentException>(), "The expanded argument constructor should validate client options");
         }
@@ -254,7 +281,21 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
         [Test]
         public void ConstructorWithSharedKeyCredentialArgumentsValidatesOptions()
         {
-            var token = new ServiceBusSharedAccessKeyCredential("key", "value");
+            var token = new AzureNamedKeyCredential("key", "value");
+            var invalidOptions = new ServiceBusClientOptions { TransportType = ServiceBusTransportType.AmqpTcp, WebProxy = Mock.Of<IWebProxy>() };
+            Assert.That(() => new ServiceBusClient("fullyQualifiedNamespace", Mock.Of<TokenCredential>(), invalidOptions), Throws.InstanceOf<ArgumentException>(), "The expanded argument constructor should validate client options");
+        }
+
+        /// <summary>
+        ///    Verifies functionality of the <see cref="ServiceBusClient" />
+        ///    constructor.
+        /// </summary>
+        ///
+        [Test]
+        public void ConstructorWithSasCredentialArgumentsValidatesOptions()
+        {
+            var signature = new SharedAccessSignature("sb://fake.thing.com", "fakeKey", "fakeValue");
+            var token = new AzureSasCredential(signature.Value);
             var invalidOptions = new ServiceBusClientOptions { TransportType = ServiceBusTransportType.AmqpTcp, WebProxy = Mock.Of<IWebProxy>() };
             Assert.That(() => new ServiceBusClient("fullyQualifiedNamespace", Mock.Of<TokenCredential>(), invalidOptions), Throws.InstanceOf<ArgumentException>(), "The expanded argument constructor should validate client options");
         }
@@ -323,7 +364,7 @@ namespace Azure.Messaging.ServiceBus.Tests.Client
             }
 
             internal ReadableOptionsMock(string fullyQualifiedNamespace,
-                                       ServiceBusSharedAccessKeyCredential credential,
+                                       AzureNamedKeyCredential credential,
                                        ServiceBusClientOptions clientOptions = default) : base(fullyQualifiedNamespace, credential, clientOptions)
             {
             }