+++ sdk/modelsrepository/Azure.IoT.ModelsRepository/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Release History
 
-## 1.0.0-preview.2 (Unreleased)
+## 1.0.0-preview.3 (Unreleased)
 
+
+## 1.0.0-preview.2 (2021-03-30)
+
+### Breaking changes
+
+- Changing the package namespace from `Azure.Iot.ModelsRepository` to `Azure.IoT.ModelsRepository`
 
 ## 1.0.0-preview.1 (2021-03-10)